# Use xxhash instead of sha256 for hashing AST nodes